refactor : global_layout dynamic 적용했습니다
