+++ src/components/layout/GlobalLayout/index.tsx
@@ -1,7 +1,10 @@
+import dynamic from "next/dynamic";
 import React from "react";
 
-import BottomNavigation from "./ui/BottomNavigation";
-import RootModal from "./ui/RootModal";
+import ServerModal from "./ui/ServerModal";
+
+const BottomNavigation = dynamic(() => import("./ui/BottomNavigation"), { ssr: false, loading: () => null });
+const ClientModal = dynamic(() => import("./ui/ClientModal"), { ssr: false, loading: () => null });
 
 type LayoutProps = {
   children: React.ReactNode;
@@ -12,7 +15,8 @@ const GlobalLayout = ({ children }: LayoutProps) => {
     <div className="mx-auto mb-14 w-full min-w-[360px] max-w-[600px] pt-14">
       {children}
       <BottomNavigation />
-      <RootModal />
+      <ClientModal />
+      <ServerModal />
     </div>
   );
 };
