--- src/components/layout/GlobalLayout/ui/RootModal/index.tsx
@@ -1,13 +1,0 @@
-import ClientModal from "./ui/ClientModal";
-import ServerModal from "./ui/ServerModal";
-
-const RootModal = () => {
-  return (
-    <>
-      <ServerModal />
-      <ClientModal />
-    </>
-  );
-};
-
-export default RootModal;
